+++ core/isiLive_db_schema.lua
@@ -206,6 +206,7 @@ local SCHEMA = {
   soundGroupJoinEnabled = { type = "boolean", default = true },
   soundLeadEnabled = { type = "boolean", default = true },
   soundPortalAvailableEnabled = { type = "boolean", default = true },
+  soundIncomingSummonLoopEnabled = { type = "boolean", default = true },
   soundReadyCheckCompleteEnabled = { type = "boolean", default = true },
   soundTankDiedEnabled = { type = "boolean", default = true },
   soundHealerDiedEnabled = { type = "boolean", default = true },